update startup bash
