+++ README.md
@@ -55,15 +55,13 @@ root@2cd33efd97b3 WORK_REPO# export DISPLAY=192.168.3.38:0.0 # change to your ow
 root@2cd33efd97b3 WORK_REPO# source venv/bin/activate
 (venv) root@2cd33efd97b3 WORK_REPO# cd /WORK_REPO/CGRA-Flow
 
-# Three themes are available: dark, light, and default
-# Start with default theme
+# Three themes are available: dark, light, and classic
+# Start with default theme (dark mode)
 (venv) root@2cd33efd97b3 build# ./launchUI.bash
-# Start with dark theme
-(venv) root@2cd33efd97b3 build# ./launchUI.bash dark
 # Start with light theme
-
-(ve
-nv) root@2cd33efd97b3 build# ./launchUI.bash light
+(venv) root@2cd33efd97b3 build# ./launchUI.bash light
+# Start with classic theme
+(venv) root@2cd33efd97b3 build# ./launchUI.bash classic
 ```
 
 Otherwise, if you don't need the GUI, development can be performed in the container with the environment well set up:
